migrations and model validations
