+++ public/components/player/edit/edit.stache
@@ -37,7 +37,9 @@
   </div>
   <button type="submit" class="btn btn-primary" >{{#player.isNew}}Create{{else}}Update{{/player.isNew}}</button>
   <a class="btn btn-default" ($click)="cancelEvent(%element)">Cancel</a>
-  {{#is playerPromise.state 'rejected'}}
-    {{playerPromise.reason.responseJSON.message}}
+  {{#is savePromise.state 'rejected'}}
+    {{#each savePromise.reason.responseJSON}}
+      <p>{{.}}</p>
+    {{/each}}
   {{/is}}
 </form>
