+++ public/components/tournament/list.stache
@@ -28,8 +28,10 @@
 	    	{($value)}="tournament.date"/>
 	</div>
 	<button type="submit" class="btn btn-default" >Create</button>
-	{{#is tournamentPromise.state 'rejected'}}
-		{{tournamentPromise.reason.responseJSON.message}}
+	{{#is savePromise.state 'rejected'}}
+		{{#each savePromise.reason.responseJSON}}
+			<p>{{.}}</p>
+		{{/each}}
 	{{/is}}
 </form>
 {{/if}}
