+++ public/components/user/create.stache
@@ -20,7 +20,9 @@
 	    
 	</div>
 	<button type="submit" {{#if user.isSaving}}disabled{{/if}} class="btn btn-default">{{#if user.isNew}}Create{{else}}Update{{/if}}</button>
-	{{#is userPromise.state 'rejected'}}
-		{{userPromise.reason.responseJSON.message}}
+	{{#is savePromise.state 'rejected'}}
+		{{#each savePromise.reason.responseJSON}}
+			<p>{{.}}</p>
+		{{/each}}
 	{{/is}}
 </form>
