+++ public/components/user/test.js
@@ -35,10 +35,11 @@ QUnit.test('createUser without password fails', function(assert) {
 
 	vm.attr('user').attr({email: 'test@bitovi.com'});
 
-	assert.expect(1);
+	assert.expect(2);
 
 	vm.createUser().fail(function(resp, type){
 		assert.equal(type, 'error', 'fail creation without password');
+		assert.equal(vm.attr('savePromise').state(), 'rejected');
 		done();
 	});
 });
